Moved signal handler into init to avoid pyflakes unused import
